[FIX] sign: filter out field type on the reference document
Before this commit, when a new template was created from a mail thread
record, the reference doc was not used to filter out the fields to
display. it would prevent to drag and drop sale.order fields
even if the template was created from the sale application.

task-5085468

Part-of: odoo/enterprise#94591
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>