[FIX] account_accountant,account_iso20022: Handle branches with end to end uuid
The aim of this commit is handling branches cases with reconciliation via the
end to end uuid.

Before this commit, a bank transaction from a company branch couldn't be
reconciled with a payment from the main company. In some situation this case
could happen. Now payments and bank transactions are reconciliable even if both
are from another company. It works only for companies with the same main company,
branches which are sisters or parent-children relation.

task-5081684

closes odoo/enterprise#94592

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>