[FIX] documents: show alias management section in details panel
Steps to Reproduce
===================
1. Set an alias in settings.
2. Go to a folder in Documents.
3. Open document right panel in list view.
-> The email alias section in the document detail panel is missing in the list view.

Issue
=======
The commit [1] unintentionally change few lines of code in forward-port conflict.
It removes `mail_alias_domain_count` from the view and we get undefined in JS.

After this commit
==================
Added required fields for the details panel to the controller mixin in JS
to make it stable friendly and simplify usage regardless of view arch.

We keep alias domain hidden when alias name is not set.

Adapt selectors of `documents kanban: select a range with SHIFT key` test as we
always pass `type` from js which groups documents into folder and file.

[1] https://github.com/odoo/enterprise/commit/d278f9013113c6de4b3425fecad5d54a9d4d22b9

Task-5045288

closes odoo/enterprise#98275

X-original-commit: c0a74324137bd9690cd75f51e5767461650a1f6b
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Mayur Raj Sinh Rathod (mrsr) <mrsr@odoo.com>