[FIX] sign: avoid mixing field models
Before this commit, it was possible to mix several field type belonging to several
models. Step to reproduce:

1) create custom sign field linked to different models:
e.g. amoint_total linked to sale order, invoice date linked to
account.move
2) Create a sign template, add these fields
3) Go to the template configuration and assign a model to this template
4) The following traceback occurs:

```
UncaughtPromiseError > TypeError

Uncaught Promise > Cannot read properties of undefined (reading 'item_type')

Occured on 88901505-19-0-all.runbot177.odoo.com on 2025-09-12 12:05:58 GMT

TypeError: Cannot read properties of undefined (reading 'item_type')
    at SignTemplateIframe.getContext (https://88901505-19-0-all.runbot177.odoo.com/web/assets/c6fae45/web.assets_web.min.js:29955:334)
    at SignTemplateIframe.renderSignItem (https://88901505-19-0-all.runbot177.odoo.com/web/assets/c6fae45/web.assets_web.min.js:29435:95)
    at SignTemplateIframe.renderSignItem (https://88901505-19-0-all.runbot177.odoo.com/web/assets/c6fae45/web.assets_web.min.js:29675:39)
    at SignTemplateIframe.renderSignItems (https://88901505-19-0-all.runbot177.odoo.com/web/assets/c6fae45/web.assets_web.min.js:29430:196)
    at SignTemplateIframe.renderSignItems (https://88901505-19-0-all.runbot177.odoo.com/web/assets/c6fae45/web.assets_web.min.js:29704:25)
    at https://88901505-19-0-all.runbot177.odoo.com/web/assets/c6fae45/web.assets_web.min.js:29421:249
    at async SignTemplateIframe.start (https://88901505-19-0-all.runbot177.odoo.com/web/assets/c6fae45/web.assets_web.min.js:29421:50)
    at async SignTemplateIframe.start (https://88901505-19-0-all.runbot177.odoo.com/web/assets/c6fae45/web.assets_web.min.js:29652:900)
```

As mixing fields from different models and linking a template to a
specific models is not a good idea, we prevent that situation to happen:
when fields linked to different models are added to a template, we
prevent the template to be linked to a single model

Note: this does not apply to res.partner fields.

task-5085468

Part-of: odoo/enterprise#94591
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>