[FIX] ai_website_livechat: protect access to classList
Steps to reproduce:
- Go to the livechat app.
- Delete all the livechat channels.
- Open the website app.
- Click on edit -> Contacts & Forms and add the ai_livechat snippet.
- Set the 'AI Agent' on the snippet and save.
- Start chatting with the ai agent using the snippet.
- An error occurs.

This error will happen in any case such that the livechat isn't available. For example:
- Deleting all the livechat channels.
- Installing a fresh db without demo data.

This will result in the element '.o-livechat-root' not appearing on
the website. This causes the ai_livechat component to have an error
when trying to hide that element. This commit fixes the issue by
adding optional chaining before accessing the class list of the element
'.o-livechat-root'.

closes odoo/enterprise#94622

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>