[FIX] stock_barcode: no longer display description when same as name
Steps to reproduce:
- Create a product
- Open the Barcode app
- Create a new reception and add the product

Issue:
The description is displayed even though it's the same as the product
name.

Since in barcode we don't get the product's `display_name` with its code
as they're treated separately, we need to reconstitute it to be sure
it's not the same as the description in order to know if we want to
display it or not.

Task-4901289

closes odoo/enterprise#94840

X-original-commit: 3ec5657c802700f68982f00968384fb2d857c0ac
Related: odoo/odoo#227348
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>