[FIX] sign: remove problematic spaces in front of paths
When trying to re-export the .pot file of the sign module, the entries
of the .pot file with a location that had a space in it, would be split
over multiple lines, making it impossible for Odoo to parse it
correctly. This commit removes the spaces in front of the paths.

Part-of: odoo/enterprise#94469
Related: odoo/odoo#226599
Related: odoo/design-themes#1145
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>