[FIX] sale_subscription: align next invoice date of postpaid
Before this commit, when a subscription whose lines were all postpaid
and the subscription plan was meant to align periods to first day of the
period (e.g. first of the month), there was an issue at confirmation.

The next invoice date was not put at the first day of the period in that
case.

closes odoo/enterprise#94746

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>