[FIX] appointment: place guests after questions
Following the addition of a phone question type (odoo/enterprise#82973),
the user phone input at registration is in the list of questions
and not at the top of the form in the base user information section anymore.
The Add Guests feature can thus be displayed before the phone number
which seems a bit odd.

Fixing the issue by moving the Add Guests feature at the end of the form,
after the list of questions.

Task-5079021

Part-of: odoo/enterprise#96027
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>