[FIX] appointment: do not send "Appointment Booked" template to
attendees

[1] introduces the new paradigm of always sending emails to "relevant
recipients" which fetches emails and partners linked to the relevant
record to send a message.

In appointment the "Appointment Booked" template is only meant to be
sent to followers of `mt{_calendar,_appointment}_event_booked` to inform
users that a new appointment was created even if they are not personally
assigned to it. Instead of setting use_default_to to False on the
template we set the (direct) recipients to False in the tracking method
so that only followers will receive it.

`test_request_meeting_message_for_manual_confirmation` is also updated
to represent the case of some visitor creating booking an appointment
instead of using internal users for everything to better represent real
use cases. Additionally each mail.mail record is extracted and checked
individually to make sure we send the right contents to the right
recipients for every recipient, as `assertMailMailWRecord` is unable to
distinguish similar mail.mail records.

[1]: 1dd6070ecaab385446cc2df7cad444f046812061

task-5075513 task-4711415

closes odoo/enterprise#94646

X-original-commit: 95fa9f75186ffb37e325da527d4e977675a429ca
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>