[FIX] l10n_{uk,nz}_reports: use fiscal year for tax return report
Steps to reproduce:
- Install Accounting and l10n_uk_reports
- Switch to a British company (e.g. UK Company)

- In Accounting settings, set the last day of the fiscal year to
another date than "December 31" (e.g. "January 31")
- Create some invoices with tax between the 1st of January and
the 30th of April

- Go to "Accounting / Accounting / Closing / Tax Returns"
- Open the "Tax Report (GB)" for the first quarter (i.e. Tax Q1)

Issue: (same issue for NZ localization)
The Tax Report is ignoring the configured date of the fiscal year.
All the invoices from January are included and those from April
are excluded.
It should be the opposite.

Solution:
Override the "_get_start_date_elements" method for the British tax
report that allows to define the start date.

opw-4939971
opw-5067877

closes odoo/enterprise#94936

X-original-commit: 12d661cb220e23ebc603efae4c9d3ac08eaf6ced
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>