[FIX] hr_contract_salary: move steps to correct submodule
`l10n_be_hr_contract_salary` adds a bunch of fields to the salary
configurator (by way of new hr.contract.salary.personal.info
records). These fields can not be filled in `hr_contract_salary` as
they are not present there, thus if the employee flow tour is run with
just `hr_contract_salary` installed it fails as soon as it tries to
fill one of these additional fields.

Move the filling of the fields to `l10n_be_hr_contract_salary`
extending the base tour (technically it might be possible to move just
the lang and remove the rest since that's the only required field).

https://runbot.odoo.com/odoo/error/232583

Part-of: odoo/enterprise#94955
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>