[FIX] l10n_dk_report: name, translation and correction of minimal report
In the minimal reports of l10n_dk, it appears that some expression ended with a
sign and the report engine was given wrong value.
This commit will remove the extra sign to correct the report and also remove the
letter or number before the name.

This commit will remove the extra sign to correct the report and also remove the
letter or number before the name.

This commit will change the translation accordingly to the other commits

task-4949062

closes odoo/enterprise#95160

X-original-commit: e9bf8ee37d9fde09a1599d64b44f0c6599299cf4
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>