[FIX] pos_self_order_preparation_display: make cancel button conditional
Currently, it is possible to cancel orders that have
been sent to the kitchen display. However when doing so,
the kitchen displayd does not receive any information
about the cancellation.

Steps to reproduce:
-------------------
- Modify restaurant and enable self ordering
- Open self an place an order (not paid but sent to kitchen)
- Go back to "My orders" and cancel it
> Order is cancelled in backend but not in the kitchen display

Why the fix:
------------
When an order is sent to the kitchen, the customer should not be
allowed to cancel it on its own. If that's the case, we now warn
the customer that their order cannot be cancelled.

opw-5030223

closes odoo/enterprise#98818

X-original-commit: a5d1021aa0533f0b4e1fe852137f96111082a2ed
Related: odoo/odoo#234408
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>