[FIX] hr_contract_salary: fix version unarchiving of new signed offer
During the offer signing process, the employee version with all his personal info is inactive and a 'fake' one is active.
On offer final signature by the applicant, the valid version is unarchived.
To avoid triggering the constraint stating that two versions cannot share the same date, we set the date of the 'fake' version to be a different date.

As the cache flush was performed after the unarchiving of the correct employee version, the constraint could not see the date changes.
This commit fixes the issue by flushing the cache before unarchiving the employee version.
Also, the second flush was unnecessary.

closes odoo/enterprise#94955

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>