[FIX] l10n_au_reports: no customer payment in TPAR report
Customer payment shoudn't be included in the TPAR report

Steps:

- Unarchive 10% TPAR tax
- Make a bill for a partner X, set 10% TPAR tax on the invoice line and
  confirm
- Create and confirm a customer payment for partner X
- Go to 'Taxable Payments Annual Reports (TPAR)'
-> The column 'Gross Paid' includes the customer payment

Before this commit, we took all line from bank or cash journal to
calculate the gross paid value.

With this commit, we also restrict the lines to include only the one
with 'asset_current' or 'liablility_current' account type.

opw-5045457

closes odoo/enterprise#95680

X-original-commit: 2fd5510d4ff3f1017d5214559ab8d0293a502c11
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>