[FIX] planning: fix planning calendar test
This commit's purpose is to fix the planning calendar print test.

Source of the issue:
the empty dict that the mockrpc returns is evaluated as a truthy value
and leads to an extra execute of the doAction function with an invalid
value.

Solution:
Return false instead of an empty dict

task-5079063

closes odoo/enterprise#95023

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>