[FIX] ai_fields: fix HTML escaped when it shouldn't
Bug
===
Since odoo/odoo@560fca7e94b232a7e2c9ad452a9772367c34204b ,
`fixInvalidHTML` escape the content if it's not a markup object.

All place have been updated, except when computing an HTML with AI.

So we make the object a markup, so it's not escaped by `fixInvalidHTML`.

The content is sanitized in `parse_ai_response`.

Standard HTML fields are not impacted, because `parseServerValue`
is called, and it transforms the string to a markup object.

Task-5092600

closes odoo/enterprise#95038

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>