[FIX] quality_control: fix traceback while click into quality points btn
Step to Reproduce -
1- Install point_of_sale and quality_control.
2- Open POS -> Product -> Product
3- Open any product and click the Quality Points smart button → traceback occurs

Root Cause -
https://github.com/odoo/odoo/blob/a729578afb7fed79aac2d622aae4da4c0917f8e5/addons/point_of_sale/views/product_view.xml#L23-L25

View reference is passed in the context.
When this context is propagated to `action_see_quality_control_point`,

https://github.com/odoo/enterprise/blob/e80b2eefd9eae99dd857fd01581c552dc83640a1/quality_control/models/quality.py#L652-L655

https://github.com/odoo/enterprise/blob/2e08282ca275bf1e66c58e28391f11f8bd7884d2/quality_control/views/quality_views.xml#L859-L868

- This happened because the action did not pass a view_id. When the context
contained list_view_ref, it tried to load the product template list view with
model quality.point, leading to a traceback since the fields from that view are
not available on the quality.point model.

Solution-

Pass a proper view_id from the Python side to ensure the correct view is loaded,
preventing `list_view_ref` from forcing an invalid template.

opw- 5090505

closes odoo/enterprise#96532

X-original-commit: 71c673a9a129acce4f8c8ad1320f498308f8ca43
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Pawan Kumar Gupta (pkgu) <pkgu@odoo.com>