[FIX] account_reports: prevent error when editing the budget report items
Currently, an error occurs when user editing the budget report items.

 - Install the Accounting module.
 - Go to Profit and Loss > Budget and create a budget.
 - Select custom dates (e.g., start: 01/01/2025, end: 12/10/2025) and
   change the amount of a budget line.
 - Change the date range (e.g., start: 01/10/2025, end: 12/10/2025) and
   change the amount again.
 - Switch back to the first date range (start: 01/01/2025, end: 12/10/2025)
   and try changing the amount once more.

`TypeError: unsupported operand type(s) for +: 'float' and 'NoneType'`

This error occurs when a user editing the budget report items. When
user enters a date period, the system creates budget items for the first date
of every month within that range. If the user then changes the date period
to the next date of the same month, the system attempts to fetch the
existing budget item [1] for that range. However, due to the start date
alignment, it fails to fetch the correct budget item and instead creates an
extra one [2]. Later, when the system checks again from the first date of
the same month as the start date, it finds this extra budget item, for that
the amount is None, which raises the error [3].

This commit ensures that when fetching existing items and generating
start month dates [4], if the start date is not the first day of the month,
it uses the first day of the next month as the start date to retrieve the
correct records.

[1]: https://github.com/odoo/enterprise/blob/1df83837a2aec4801b34a9a7ab0cd68f640b4fd6/account_reports/models/budget.py#L44-L49

[2]:
https://github.com/odoo/enterprise/blob/1df83837a2aec4801b34a9a7ab0cd68f640b4fd6/account_reports/models/budget.py#L75-L79

[3]: https://github.com/odoo/enterprise/blob/1df83837a2aec4801b34a9a7ab0cd68f640b4fd6/account_reports/models/budget.py#L72

[4]: https://github.com/odoo/enterprise/blob/1df83837a2aec4801b34a9a7ab0cd68f640b4fd6/account_reports/models/budget.py#L58-L61

sentry-6883207225

closes odoo/enterprise#95976

X-original-commit: 721182eee549f110b8e992fa17e9a98e5ea41396
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>