[IMP] l10n_br_edi_fiscal_reform*: 2026 fiscal reform
2026 marks the start of Brazil's 8-year plan to migrate to a
completely new consumption tax system [1]. The new system is meant to
be simpler than the old one, but we defer to Avalara for tax
calculation so we don't really notice it technically.

The old and new system have to coexist for the foreseeable future
because the rollout happens gradually over many years. Whether the new
system has to be enabled depends on the federal state the business is
in, the type of transactions they do, and other factors. The user has
the option to enable it through the l10n_br_is_icbs Boolean flag on
the company, which gates all the new functionality.

These changes impact both tax calculation and EDI. This set of modules
adapts our existing l10n_br* modules to the new requirements set out
by Avalara. We are required to send extra values to their API, and in
some cases adapt existing ones.

task-4939796

[1] https://www.planalto.gov.br/ccivil_03/leis/lcp/lcp214.htm

closes odoo/enterprise#102836

X-original-commit: 3908a9e48fd8e4cf1cf07cc34bbb15eb64271c9c
Related: odoo/odoo#241221
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>