[FIX] l10n_ec_edi_pos: prevent error when validating POS order
After the commit https://github.com/odoo/enterprise/commit/343efe41bc74e612c867d5109c196a11680866db, the `currentOrder` reference in the
payment validation flow was replaced with `order`. However, this change
was not reflected in `l10n_ec_edi_pos`, which still relied on
`currentOrder`.

As a result, attempting to validate a POS order raised an error, making
it impossible to complete the order.

This commit updates the logic to use `order` instead of `currentOrder`,
aligning with the upstream changes and restoring proper functionality.

opw-5099083

closes odoo/enterprise#95105

Signed-off-by: David Monnom (moda) <moda@odoo.com>