[IMP] account_accountant: auto reco model name
In this commit: https://github.com/odoo/enterprise/commit/9cdc519a7aa7e40cad31d2eb6dcc8858df831e9b
we change the way the reco model name were made by going back to just the
display name of the account. For prior version 19.0, a unique constraint on the
name was there so we added a number after the name of the account to avoid
duplicate. The constraint has been removed and so this code is no longer needed.

no task id

closes odoo/enterprise#95130

Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>