[FIX] appointment: fix max capacity on refresh
When refreshing the slots, it's possible that the resource_selected_id
is equal to None, False or just empty string. This was leading to some
error when parsing it to an integer.
This commit move the parsing into the method computing the max possible
capacity after checking if we got a value.

Related commit 3cca7e47ab58f8a7d4e9196dbf60f7068348216b

task-5102895

closes odoo/enterprise#95380

X-original-commit: 53fd5990d6b5e71371d1edd6e3837046690c9178
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>