[FIX] account_accountant: mobile view traceback reconcile button
When clicking on the reconcile button in mobile, the js_class of the custom
list view is never triggered because, in small screens (aka. "mobile"), by
default, kanban views are used instead of list views (or at least take
precedence).

To avoid having a traceback, we will keep the custom behavior in desktop only

no task id

closes odoo/enterprise#95499

X-original-commit: 97efa69bf1942e10f727e15ca5af39f25aa0690f
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>