[FIX] helpdesk_timesheet: include helpdesk ticket in timesheet report
to reproduce:
=============
1. make helpdesk team billable and records timesheets
2. create a helpdesk ticket and link it to a sale order
3. log timesheets on the ticket
4. print the timesheet report **from the sale order**
-> the task column will contain only the helpdesk team name, while it
should contain "helpdesk team / ticket name"

Problem:
========
in helpdesk_timesheet we inherit `hr_timesheet.timesheet_table` to adapt
it to helpdesk tickets, but we use `show_ticket` to display the ticket
name, which is only set in `hr_timesheet.report_timesheet` and
`hr_timesheet.timesheet_project_task_page` but not in `sale_timesheet.timesheet_sale_page`
which is the one used when printing the report from the sale order.

Solution:
=========
`show_ticket` should be set with value `bool(lines.helpdesk_ticket_id)`
which is equivalent to `line.helpdesk_ticket_id` in the t-if condition.
so we can directly use `line.helpdesk_ticket_id` and remove the
`show_ticket` variable.

opw-5002650

closes odoo/enterprise#95239

X-original-commit: 9c972fca4fb8f7a930680765ce424c7fa667b52a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>