[FIX] sale_renting_planning: add order line on adding shift to last rental order
steps to reproduce:
- Configure a booking product as service and rental.
- Confirm a shift to create a rental order.
- Activate sync on the role.
- Add a new shift with the same product using "Add to last order".
- Open the rental order: the order line and total price are not updated.

issue:
The price and the order line were not updated when a new shift was added to the
last order.

cause:
Previously, adding a new shift to a rental order did not add the order
line on the corresponding order. becacuse we was not checking the state of the
order and at line creation line will be created for confirmed order only.

fix:
A check is added when the fetching the rental order it should be confirmed, and
adding shift to last order will increase qyt of the line.

task-5093190

closes odoo/enterprise#100878

X-original-commit: c394f00c6edd5890cbc9f20ea79ee6399bbfe188
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>