[FIX] sign: prevent crash when multiple line item are constant
When a multi line element is used with a constant value, the following
traceback is observed:

error_service.js:87 TypeError: Cannot read properties of undefined (reading 'split')
    at SignablePDFIframe.textareaApplyLineBreak (:8069/web/assets/deb…e_PDF_iframe.js:536)
    at Object.textarea (:8069/web/assets/deb…e_PDF_iframe.js:501)
    at SignablePDFIframe.getSignatureValueFromElement (:8069/web/assets/deb…e_PDF_iframe.js:522)
    at SignablePDFIframe.getSignatureValuesFromConfiguration (:8069/web/assets/deb…e_PDF_iframe.js:453)
    at Document.signDocuments (:8069/web/assets/deb…ent_signable.js:386)
    at HTMLButtonElement.<anonymous> (:8069/web/assets/deb…ent_signable.js:287)

The reason is that we call textareaApplyLineBreak on item.el but
item.el.value does not exists for constant item.
As a result, it crashes. We decided to fix it by making sure we call
this function only if the element is not constant to avoid trigger an
error where we try to write on a constant value that should not be
updated.

task-5104647

Part-of: odoo/enterprise#95196
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>