[FIX] l10n_mx_edi: Create XML as Superuser
When creating an XML attachment as a user without Write access on the
ir.ui.view model, the Mimetype will be set to plain/text. In particular,
this causes issues when Accounting centralization is enabled in
Documents, as the corresponding Document will only be generated if the
Mimetype is application/xml.

Creating the XML as Superuser avoids this issue.

Similar to https://github.com/odoo/odoo/pull/124507

opw-5057038

closes odoo/enterprise#97325

X-original-commit: 313fd7f499b09df37674f266c0d2be2f9c140566
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>