[FIX] helpdesk_timesheet: correct previous report <td> update for tickets
Description of the issue/feature this PR addresses:
The previous commit attempted to extend the timesheet report <td> to display tickets by using position="attributes" on a new <td>. This approach does not work in Odoo reports because position="attributes" can only modify existing elements. There is no indication that the behavior of not displaying tickets was intentional, so this PR corrects that implementation.

Current behavior before PR:
- The previous fix did not correctly locate the existing <td> element for task/project info.

Desired behavior after PR is merged:
- The existing <td> is correctly found and updated to include show_ticket in its t-if.

closes odoo/enterprise#95341

X-original-commit: 6d89b2aa3922ff6ea2e0000fd0359c4dcf35f031
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>