[IMP] l10n_mx_*: Change demo company.
The SAT has a list of ‘demo’ companies to enable invoicing demo operations through Electronic Data Interchange (EDI). Currently, in Odoo, the default demo company is Kemper School. However, with the release of payroll stamping in version 19, this company is no longer useful, as it cannot stamp payroll. So, instead of adding another company, we modified the Kemper School data to change it to another company that can stamp payroll and all other existing documents.

The demo company data was changed, the certificates were changed, and a file that was never added to the manifest and therefore had no use was deleted.

I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#95221

Related: odoo/odoo#228048
Signed-off-by: Laurent Smet (las) <las@odoo.com>