[IMP] account_accountant: show analytic distribution warning
When setting the account on a bank statement line,
analytic distribution might be required due to a mandatory plan.

Before this commit, the user had no idea if the analytic is
required or not.

After this commit, lines that required analytic distribution
will have a warning directing the user to set it.

task-5055351

closes odoo/enterprise#95266

closes odoo/enterprise#100798

Related: odoo/odoo#228110
Related: odoo/odoo#237837
X-original-commit: 74a44f942da21000368e97a3487fe2cfb937c06f
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>