[IMP] account_accountant: tolerance of matching
This commit: https://github.com/odoo/enterprise/commit/9bea3f2c517e77fdb822eb53b9648e9fc5478dac
introduce a tolerance. If an invoice has a bigger amount than the bank line but
within the tolerance of 3%, reconcile the invoice automatically and leave
the difference in the suspense account.
The statement line will be partially matched.

This commit will change that tolerance to use a system parameter instead.
For old DB the tolerance is at 0 and for new DB a system parameter (0 by
default) will be put so that people that want to use it can.

task-5106655

closes odoo/enterprise#95582

X-original-commit: 0847d4556dc71bb4a83bce45c0a16b0f0d278e31
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>