[FIX] pos_*: display order on prep display after payment confirmation
pos_*= pos_online_payment_self_order_preparation_display

Before this commit, if an online payment method was assigned to a self or a kiosk, the order was displayed on the preparation display before the payment was confirmed.

After this commit, the order is no longer displayed on the preparation display until the payment confirmation.

Backport of https://github.com/odoo/enterprise/pull/87173, with additional logic to ensure the order is correctly sent to the preparation display even if an exit route is used in the payment portal

closes odoo/enterprise#95675

X-original-commit: d0bcd920675f36223d68c6e64ad496cb266f2e70
Related: odoo/odoo#228876
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>