[FIX] account_reports: Fixed ugly partner ledger on smaller screens
Problem:
On smaller screens (phone or tablet) the center alignment between the partner name and the little arrow is broken.

Cause:
This happened because the little arrow has btn class, which gets larger on smaller screens
(to make it easier to click). The larger button causes the center alignment to be lost,
as it grows alone inside the table row.

Fix:
This commit fixes this issue by ensuring that the content of line_name and line_cell
are vertically aligned to the middle, and that items in their sub-containers are aligned to the center.

task-5106889

closes odoo/enterprise#95816

X-original-commit: 79b25c880714d14cbc183a0428919901e5b5d566
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>