[FIX] ai: execute non "code" server actions when triggered manually
Purpose:
-------
Currently, when an AI server action is triggered manually, the child server
actions with state != code are not executed.
Nothing happens because inside `_run`, the server action is executed on
each record obtained by browsing the ids given by the `active_ids` and
`active_id` keys in the context, which are not present (it works fine when
the action is triggered by an automation because the record info is added
in the context in `_process` in base automation).

This commits adds the current records info in the context of the ai tools
so that these server actions (such as create record, create activity or
send mail) are now executed).

Task-5107876

Part-of: odoo/enterprise#95343
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>