[FIX] account_accountant: fix bank rec widget tour
The step checks the non presence of `o_bank_rec_quick_create`.
But this element does not exist, as it should be `o_bank_reconciliation_quick_create`.

This causes the tour to sometimes fail as we click on unfold before the
interface updates from the creation of the line.

runbot-error-230727

closes odoo/enterprise#96269

X-original-commit: de76dde733d6b378490817b7795d7a2b3bca0ebd
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>