[IMP] voip: move session-related methods from UserAgent to Session
The original VoIP code was written under the assumption that there would
only ever be one session at a time. Consequently, much of the session
state was stored in a global object called the UserAgent.

However, Odoo 19 implemented an attended transfer in VoIP, creating
situations in which two sessions occur simultaneously. For this reason,
the asynchronous code should no longer update session properties stored
on the global UserAgent, as the session may have changed in the
meantime, which could lead to race conditions.

This commit moves some logic from the UserAgent to the Session, and
ensures all callbacks are properly mapped to the relevant session.

closes odoo/enterprise#95417

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>