[FIX] l10n_mx_reports: Fix columns positions in Diot.
The description of the columns “exempt imports” and “exempt” is somewhat ambiguous, so when developing the diot, there was a small error in the order of the columns. The values are calculated correctly, but columns 49 and 50 were reversed.

This Pr changes the order of the columns in the downloadable file (txt) and corrects the tests due to the change.

Task-id: 5096808

I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#95711

X-original-commit: 0d5a2f2c55ffd468c19130d9a99871d2734503a0
Related: odoo/odoo#228938
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Hector Mauricio Perez Barajas (hmpb) <hmpb@odoo.com>