[ADD] l10n_ar_reports_simple: add iva simple export
The Argentinian government has released support for a new type of a tax
report that companies can upload called IVA Simple. This is a zip of CSV
files that can be imported to the government portal. As such we want to
support exporting this report for our customers.

This commit adds a new auto installed module for a new report export type
of IVA Simple on the Argentinian Tax Report as well as the model data
required for the export.

task-5027104

closes odoo/enterprise#93871

Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Co-authored-by: Julia Elizondo <jue@adhoc.com.ar>