[FIX] web_studio: fix monetary related creation
Before this commit, when creating a related field to a monetary, the created currency field
was not stored and not related either, so the monetary was unusable.

This was because of 5cf5a35a0a8f78655989009d0eddcf39f8430965 , b177b058be1531c3d2af2b591c22591c19240d33
and in general the changes in read_group that largely improve the situation.

After this commit, we create a currency field related to the currency field of the related monetary
to ensure that the monetary's value is coherent.

This is made possible by the above mentionned improvements in read_group

opw-5094619

closes odoo/enterprise#95810

X-original-commit: c57c247490f0f58c8746af47ce035c28137faf97
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>