[IMP] hr_attendance: adding overtime ruleset data
It's impossible to create an attendance for an employee without ruleset.
To solve this problem, this PR adds a default ruleset with 2 rules: one for
overtime, and another one for non working days.

Those data are the same as the data added in the
upgrade/migrations/hr_attendance/saas~18.5.2.0 post-migrate script.

task-5082628

closes odoo/enterprise#95421

Related: odoo/odoo#228352
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>