[FIX] web_studio: edit_view: remove attribute with None value
On a calendar with studio, try to remove the "color" attribute, or
any other that should contain the name of a field.

Before this commit there was a crash because the value sent to the server in this case
is `undefined` (`null` in JSON or `None` in python), which was stringified and yielding
an actual string that was not a field name

After this commit, NULL values are not stringified, instead they should represent
the emptiness of the attribute.

opw-4938351

closes odoo/enterprise#96214

X-original-commit: a094d446b03da898ac1b5a8e26648bbca81fa81f
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>