[IMP] l10n_se_bban: Allow batch payment with iban and bban
This commit add a new behaviour for batch payments in Sweden.

Before this commit: batch payment with mixed payment for iban and bban/bankgiro/plusgiro wasn't allowed

After this commit: now if the user create a batch payment with both type of account, we create a zip file with 2 xml files, one for iban
and one for bban/bankgiro/plusgiro.

task-5107240

X-original-commit: 7bd7661bbb30debf31c3f0d49a81f7af8f343639
Part-of: odoo/enterprise#97648
Related: odoo/odoo#232318
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>