[IMP] ai_documents: open server actions on click in ai sort modal
Purpose:
--------
When editing the "auto-sort" prompt on a folder, the tools that the AI
can take can now be opened by clicking on them.

The "default_use_in_ai" key has also been added in the context so that
server actions created from this modal are usable in AI by default, as
they are intended to be used by the ai auto-sort action.

Task-5107876

closes odoo/enterprise#95343

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>