[IMP] account_reports: Fetching annotations based on create date when exporting
Previously when exporting a report, we would sort the annotation based
on their create_date or to get the last annotation we would need to
use max to find the latest one.

Now, when exporting the report we order them directly so we dont need to
bother with it. It's not needed to sort them when simply viewing the
report as they are displayed in the chatter already in the right order

Also fix a bug where the last_message displayed on the balance view of
the audits isn't filtered on the period of the audit.
To replicate:
- Create an Audit for any year.
- Post an annotation on an account on a period outside of the period
  of the audit.
- Go to the balance view and check the account, it now has the new
  message even though it's outside the audit period.

closes odoo/enterprise#95307

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>