[FIX] whatsapp: message translation for agent when not member
Before this commit, when an agent that is not member of whatsapp but
is peeking the conversation, the agent could not translate the
message.

This happens because the translation feature is limited to internal
users, but this was determined based on the self member relational
field. This works when the agent is a member but when not a member
this was arbitrarily disabling the feature.

This commit fixes the issue by looking at whether the user is
internal or not, based on self persona independently on whether the
agent is member or not of the conversation.

Task-5111383

closes odoo/enterprise#96205

X-original-commit: db09bc80ed657868f2c0e935af330a61acb7f5ff
Related: odoo/odoo#229804
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>